requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
gunicorn==23.0.0
torch==2.8.0
torchvision==0.23.0
timm==1.0.19
Pillow==11.3.0
scikit-learn
PyYAML
numpy